analytics/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pymongo==4.9.1
python-dotenv==1.0.1
